Move mocks to tests dir
